+++ tests/data/cases/pep_750.py
@@ -40,6 +40,11 @@ t'''
     {1}_cte AS ()'''}
 '''
 
+# Uppercase T prefix is normalized to lowercase t
+x = T"bar {1 + 1}"
+x = T'bar {1 + 1}'
+rT'\{{\}}'
+
 # output
 x = t"foo"
 x = t"foo {{ {2 + 2}bar {{ baz"
@@ -81,3 +86,8 @@ t"""
     WITH {f'''
     {1}_cte AS ()'''}
 """
+
+# Uppercase T prefix is normalized to lowercase t
+x = t"bar {1 + 1}"
+x = t"bar {1 + 1}"
+rt"\{{\}}"